update view model in comply with design changes
