+++ app/src/main/java/com/doublesp/coherence/actions/IdeaCreationActionHandler.java
@@ -1,7 +1,5 @@
 package com.doublesp.coherence.actions;
 
-import android.text.Editable;
-
 import com.doublesp.coherence.interfaces.domain.IdeaInteractorInterface;
 import com.doublesp.coherence.interfaces.presentation.IdeaActionHandlerInterface;
 import com.doublesp.coherence.viewmodels.Idea;
@@ -19,23 +17,6 @@ public class IdeaCreationActionHandler implements IdeaActionHandlerInterface {
             IdeaInteractorInterface ideaInteractor) {
         mPreviewHandler = previewHandler;
         mIdeaInteractor = ideaInteractor;
-    }
-
-    @Override
-    public void afterTextChanged(Editable s) {
-        final int i = s.length();
-        if (i == 0) {
-            return;
-        }
-        if (s.subSequence(i - 1, i).toString().equals("\n")) {
-//            mIdeaInteractor.addIdea(s.toString().trim());
-            // TODO: support arbitrary user input, e.g. As an user, I want to be able to add
-            // other ingredients
-            // to the shopping list that is not listed on the recipe.
-            s.clear();
-        } else {
-            mIdeaInteractor.getSuggestions(s.toString().trim());
-        }
     }
 
     @Override
