+++ app/src/main/java/com/doublesp/coherence/actions/ListFragmentActionHandler.java
@@ -1,16 +1,18 @@
 package com.doublesp.coherence.actions;
 
-import android.content.Context;
-import android.content.Intent;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ServerValue;
 
 import com.doublesp.coherence.interfaces.domain.IdeaInteractorInterface;
 import com.doublesp.coherence.interfaces.presentation.ListFragmentActionHandlerInterface;
 import com.doublesp.coherence.utils.ConstantsAndUtils;
 import com.doublesp.coherence.viewmodels.Plan;
 import com.doublesp.coherence.viewmodels.UserList;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ServerValue;
+
+import android.content.Context;
+import android.content.Intent;
+import android.text.Editable;
 
 import java.util.HashMap;
 
@@ -66,6 +68,21 @@ public class ListFragmentActionHandler implements ListFragmentActionHandlerInter
         // with app link
         shareIntent.putExtra(Intent.EXTRA_TEXT, sharableContentBuilder.toString());
         mShareHandler.share(shareIntent);
+    }
+
+    @Override
+    public void afterTextChanged(Editable s) {
+        final int i = s.length();
+        if (i == 0) {
+            return;
+        }
+        if (s.subSequence(i - 1, i).toString().equals("\n")) {
+            mIdeaInteractor.addIdea(s.toString().trim());
+            s.clear();
+        } else {
+            // TODO: auto-complete
+//            mIdeaInteractor.getSuggestions(s.toString().trim());
+        }
     }
 
     public interface IdeaShareHandlerInterface {
