+++ app/src/main/java/com/doublesp/coherence/adapters/IdeaReviewArrayAdapter.java
@@ -1,20 +1,19 @@
 package com.doublesp.coherence.adapters;
 
-import static com.raizlabs.android.dbflow.config.FlowManager.getContext;
-
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
 import com.doublesp.coherence.R;
 import com.doublesp.coherence.interfaces.presentation.IdeaViewHolderInterface;
 import com.doublesp.coherence.viewholders.IdeaViewHolder;
 import com.doublesp.coherence.viewmodels.Idea;
 import com.doublesp.coherence.viewmodels.Plan;
 
-import java.util.ArrayList;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
 import java.util.List;
+
+import static com.raizlabs.android.dbflow.config.FlowManager.getContext;
 
 /**
  * Created by pinyaoting on 11/21/16.
@@ -25,12 +24,7 @@ public class IdeaReviewArrayAdapter extends RecyclerView.Adapter {
     List<Idea> mIdeas;
 
     public IdeaReviewArrayAdapter(Plan plan) {
-        mIdeas = new ArrayList<>();
-        for (Idea idea : plan.getIdeas()) {
-            for (Idea relatedIdea : idea.getRelatedIdeas()) {
-                mIdeas.add(relatedIdea);
-            }
-        }
+        mIdeas = plan.getIdeas();
     }
 
     @Override
