+++ app/src/main/java/com/doublesp/coherence/interactors/MockRecipeInteractor.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class MockRecipeInteractor extends RecipeInteractor {
 
     public MockRecipeInteractor(IdeaDataStoreInterface ideaDataStore,
-            RecipeRepositoryInterface recipeRepository) {
+                                RecipeRepositoryInterface recipeRepository) {
         super(ideaDataStore, recipeRepository);
         ideaDataStore.setSnapshot(Parcels.wrap(new IdeaSnapshotStore(mockIdeas())));
     }
@@ -39,8 +39,7 @@ public class MockRecipeInteractor extends RecipeInteractor {
                 false,
                 R.id.idea_type_user_generated,
                 new IdeaMeta("http://www.specialtyproduce.com/sppics/2015.png", "red potatoes",
-                        "3 lbs. small red potatoes halved"),
-                null));
+                        "3 lbs. small red potatoes halved")));
         ideas.add(new Idea("https://www.edamam.com/#recipe_3da57f2d5580d86d21b6cb569d1974f9",
                 R.id.idea_category_recipe,
                 "olive oil",
@@ -50,8 +49,7 @@ public class MockRecipeInteractor extends RecipeInteractor {
                         "http://www.medicalnewstoday"
                                 + ".com/content/images/articles/266/266258/olive-oil-and-olives"
                                 + ".jpg",
-                        "olive oil", "3 tablespoons olive oil"),
-                null));
+                        "olive oil", "3 tablespoons olive oil")));
         ideas.add(new Idea("https://www.edamam.com/#recipe_b606ed3695ad9202c0342d816a236196",
                 R.id.idea_category_recipe,
                 "ground pepper",
@@ -60,8 +58,7 @@ public class MockRecipeInteractor extends RecipeInteractor {
                 new IdeaMeta(
                         "http://rainydayfoods"
                                 + ".com/media/catalog/product/cache/1/image/800x800/9df78eab33525d08d6e5fb8d27136e95/p/e/pepper_1.jpg",
-                        "ground pepper", "Coarse salt and ground pepper"),
-                null));
+                        "ground pepper", "Coarse salt and ground pepper")));
         return ideas;
     }
