+++ app/src/main/java/com/doublesp/coherence/interfaces/presentation/ListFragmentActionHandlerInterface.java
@@ -1,5 +1,7 @@
 package com.doublesp.coherence.interfaces.presentation;
 
+import android.text.Editable;
+
 /**
  * Created by pinyaoting on 11/13/16.
  */
@@ -7,5 +9,6 @@ package com.doublesp.coherence.interfaces.presentation;
 public interface ListFragmentActionHandlerInterface {
 
     void onFloatingAcitonButtonClick();
+    void afterTextChanged(Editable s);
 
 }
