+++ app/src/main/java/com/doublesp/coherence/viewmodels/Idea.java
@@ -4,8 +4,6 @@ import com.doublesp.coherence.R;
 
 import org.parceler.Parcel;
 
-import java.util.List;
-
 /**
  * Created by pinyaoting on 11/10/16.
  */
@@ -19,20 +17,18 @@ public class Idea {
     boolean crossedOut;
     int type; // R.id.idea_type_user_generated or R.id.idea_type_suggestion
     IdeaMeta meta;
-    List<Idea> relatedIdeas;
 
     public Idea() {
     }
 
     public Idea(String id, int category, String content, boolean crossedOut, int type,
-            IdeaMeta meta, List<Idea> relatedIdeas) {
+            IdeaMeta meta) {
         this.id = id;
         this.category = category;
         this.content = content;
         this.crossedOut = crossedOut;
         this.type = type;
         this.meta = meta;
-        this.relatedIdeas = relatedIdeas;
     }
 
     public static Idea newInstanceOfBlankIdea() {
@@ -63,10 +59,6 @@ public class Idea {
 
     public IdeaMeta getMeta() {
         return meta;
-    }
-
-    public List<Idea> getRelatedIdeas() {
-        return relatedIdeas;
     }
 
     @Override
