+++ app/src/main/res/layout/fragment_list_composition.xml
@@ -1,30 +1,53 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+        xmlns:tools="http://schemas.android.com/tools">
 
     <data>
 
-        <variable name="handler"
-            type="com.doublesp.coherence.interfaces.presentation.ListFragmentActionHandlerInterface" />
+        <variable
+                name="handler"
+                type="com.doublesp.coherence.interfaces.presentation.ListFragmentActionHandlerInterface"/>
     </data>
 
     <FrameLayout xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent" android:layout_height="match_parent"
-        tools:context="com.doublesp.coherence.fragments.ListCompositionFragment">
+                 android:layout_width="match_parent"
+                 android:layout_height="match_parent"
+                 tools:context="com.doublesp.coherence.fragments.ListCompositionFragment">
 
-        <ImageView android:id="@+id/ivIdeaCompositionBackground" android:layout_width="match_parent"
-            android:layout_height="match_parent" android:alpha="0.2" android:scaleType="centerCrop"
-            tools:src="@drawable/background_0" />
+        <ImageView
+                android:id="@+id/ivIdeaCompositionBackground"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:alpha="0.1"
+                android:scaleType="centerCrop"
+                tools:src="@drawable/background_0"/>
 
-        <android.support.v7.widget.RecyclerView android:id="@+id/rvIdeas"
-            android:layout_width="match_parent" android:layout_height="match_parent"
-            android:background="@android:color/transparent" />
+        <android.support.v7.widget.RecyclerView
+                android:id="@+id/rvIdeas"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:background="@android:color/transparent"/>
 
-        <android.support.design.widget.FloatingActionButton android:layout_width="wrap_content"
-            android:layout_height="wrap_content" android:layout_gravity="bottom|right|end"
-            android:layout_margin="16dp"
-            android:onClick="@{() -> handler.onFloatingAcitonButtonClick()}"
-            android:src="@drawable/ic_share" android:tint="@color/colorButton" app:fabSize="normal"
-            app:layout_anchor="@id/rvIdeas" app:layout_anchorGravity="bottom|right|end" />
+        <EditText
+                android:id="@+id/etIdea"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/activity_list_composition_idea_height"
+                android:layout_gravity="bottom"
+                android:gravity="center"
+                android:hint="@string/idea_hint_ingredient"
+                android:maxLines="1"
+                android:textColor="@color/colorPrimary"/>
+
+        <android.support.design.widget.FloatingActionButton
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="bottom|right|end"
+                android:layout_margin="16dp"
+                android:onClick="@{() -> handler.onFloatingAcitonButtonClick()}"
+                android:src="@drawable/ic_share"
+                android:tint="@color/colorButton"
+                app:fabSize="normal"
+                app:layout_anchor="@id/rvIdeas"
+                app:layout_anchorGravity="bottom|right|end"/>
 
     </FrameLayout>
 </layout>
