--- app/src/main/res/layout/item_idea_blank.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-
-        <variable name="pos" type="int" />
-
-        <variable name="handler"
-            type="com.doublesp.coherence.interfaces.presentation.IdeaActionHandlerInterface" />
-
-        <variable name="viewModel" type="com.doublesp.coherence.viewmodels.Idea" />
-    </data>
-
-    <LinearLayout android:layout_width="match_parent" android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <EditText android:id="@+id/etIdea" android:layout_width="match_parent"
-            android:layout_height="@dimen/activity_list_composition_idea_height"
-            android:gravity="center" android:hint="@string/idea_hint_recipe" android:maxLines="1"
-            android:text='@{viewModel.content}' android:textColor="@color/colorPrimary" />
-    </LinearLayout>
-</layout>
